+++ lib/pry/pager.rb
@@ -3,16 +3,24 @@ require 'pry/terminal'
 # A pager is an `IO`-like object that accepts text and either prints it
 # immediately, prints it one page at a time, or streams it to an external
 # program to print one page at a time.
-module Pry::Pager
+class Pry::Pager
   class StopPaging < StandardError
+  end
+
+  attr_reader :_pry_
+
+  def initialize(_pry_)
+    @_pry_ = _pry_
   end
 
   # Send the given text through the best available pager (if `Pry.config.pager` is
   # enabled).
+  # If you want to send text through in chunks as you generate it, use `open` to
+  # get a writable object instead.
   # @param [String] text A piece of text to run through a pager.
   # @param [IO] output (`$stdout`) An object to send output to.
-  def self.page(text, output = $stdout)
-    with_pager(output) do |pager|
+  def page(text)
+    open do |pager|
       pager << text
     end
   end
@@ -20,28 +28,33 @@ module Pry::Pager
   # Yields a pager object (`NullPager`, `SimplePager`, or `SystemPager`).  All
   # pagers accept output with `#puts`, `#print`, `#write`, and `#<<`.
   # @param [IO] output (`$stdout`) An object to send output to.
-  def self.with_pager(output = $stdout)
-    pager = best_available(output)
+  def open
+    pager = best_available
     yield pager
   rescue StopPaging
   ensure
     pager.close if pager
   end
 
+  private
+
+  attr_reader :output
+  def enabled?; !!@enabled; end
+
   # Return an instance of the "best" available pager class -- `SystemPager` if
   # possible, `SimplePager` if `SystemPager` isn't available, and `NullPager`
   # if the user has disabled paging. All pagers accept output with `#puts`,
   # `#print`, `#write`, and `#<<`. You must call `#close` when you're done
   # writing output to a pager, and you must rescue `Pry::Pager::StopPaging`.
-  # These requirements can be avoided by using `.with_pager` instead.
+  # These requirements can be avoided by using `.open` instead.
   # @param [#<<] output ($stdout) An object to send output to.
-  def self.best_available(output)
-    if !Pry.config.pager
-      NullPager.new(output)
+  def best_available
+    if !_pry_.config.pager
+      NullPager.new(_pry_.config.output)
     elsif !SystemPager.available? || Pry::Helpers::BaseHelpers.jruby?
-      SimplePager.new(output)
+      SimplePager.new(_pry_.config.output)
     else
-      SystemPager.new(output)
+      SystemPager.new(_pry_.config.output)
     end
   end